+++ Sources/KinescopeSDK/Protocols/KinescopeConfigurable.swift
@@ -10,9 +10,12 @@ public protocol KinescopeConfigurable {
 
     /// - parameter config: Configuration parameters required to connection
     func setConfig(_ config: KinescopeConfig)
+    
+    /// - parameter delegate: Delegate to listen for analytics sended from ``KinescopeVideoPlayer``
+    func setAnalytics(delegate: KinescopeAnalyticsDelegate)
 
     /// - parameter logger: opportunity to set custom logger
     /// - parameter levels: levels of logging
     func set(logger: KinescopeLogging, levels: [KinescopeLoggingLevel])
-    
+
 }